pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["author"]

[tool.poetry.dependencies]
python = "3.8"
django = "^3.0.7"
django-filter = "^2.3.0"
psycopg2-binary = "^2.8.5"
django-environ = "^0.4.5"
djangorestframework = "^3.11.0"
pytest-django = "^3.9.0"

[tool.poetry.dev-dependencies]
pytest = "^3.4"
black = "^19.10b0"
isort = "^4.3.21"
flake8 = "^3.8.3"
